Remove old navigation and app implementation
