--- cmp-common/src/commonMain/kotlin/ru/beryukhov/coffeegram/pages/CoffeeListPage.kt
@@ -1,85 +1,0 @@
-package ru.beryukhov.coffeegram.pages
-
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import com.slapps.cupertino.adaptive.AdaptiveIconButton
-import com.slapps.cupertino.adaptive.AdaptiveTopAppBar
-import kotlinx.datetime.LocalDate
-import ru.beryukhov.coffeegram.data.CoffeeType
-import ru.beryukhov.coffeegram.data.DayCoffee
-import ru.beryukhov.coffeegram.data.coffeeTypeValues
-import ru.beryukhov.coffeegram.model.CoffeeTypeWithCount
-import ru.beryukhov.coffeegram.model.DaysCoffeesIntent
-import ru.beryukhov.coffeegram.model.DaysCoffeesState
-import ru.beryukhov.coffeegram.model.DaysCoffeesStore
-import ru.beryukhov.coffeegram.model.NavigationIntent
-import ru.beryukhov.coffeegram.model.NavigationState
-import ru.beryukhov.coffeegram.model.NavigationStore
-import ru.beryukhov.coffeegram.view.CoffeeTypeItem
-
-@Composable
-fun CoffeeListAppBar(navigationStore: NavigationStore, modifier: Modifier = Modifier) {
-    AdaptiveTopAppBar(
-        title = { Text("Add drink") },
-        navigationIcon = {
-            AdaptiveIconButton(onClick = { navigationStore.newIntent(NavigationIntent.ReturnToTablePage) }) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                    contentDescription = ""
-                )
-            }
-        }
-    )
-}
-
-@Composable
-fun CoffeeListPage(daysCoffeesStore: DaysCoffeesStore, navigationState: NavigationState.CoffeeListPage) {
-    CoffeeList(
-        navigationState.date,
-        daysCoffeesStore
-    )
-}
-
-@Composable
-fun CoffeeList(
-    localDate: LocalDate,
-    daysCoffeesStore: DaysCoffeesStore,
-    modifier: Modifier = Modifier
-) {
-    val dayCoffeeState: DaysCoffeesState by daysCoffeesStore.state.collectAsState()
-    val dayCoffee = dayCoffeeState.coffees[localDate] ?: DayCoffee()
-    LazyColumn(modifier = modifier.fillMaxHeight()) {
-        itemsIndexed(
-            items = dayCoffee.coffeeCountMap.withEmpty(),
-            itemContent = { _, (coffee, count): CoffeeTypeWithCount ->
-                CoffeeTypeItem(
-                    coffeeType = coffee,
-                    count = count,
-                    onIncrement = { daysCoffeesStore.newIntent(DaysCoffeesIntent.PlusCoffee(localDate, coffee)) },
-                    onDecrement = { daysCoffeesStore.newIntent(DaysCoffeesIntent.MinusCoffee(localDate, coffee)) }
-                )
-            }
-        )
-    }
-}
-
-class MutablePair(val ct: CoffeeType, var count: Int)
-
-// @VisibleForTesting
-internal fun Map<CoffeeType, Int>.withEmpty(): List<CoffeeTypeWithCount> {
-    val emptyList: MutableList<MutablePair> =
-        coffeeTypeValues().toList().map { MutablePair(it, 0) }.toMutableList()
-    this.forEach { entry: Map.Entry<CoffeeType, Int> ->
-        emptyList.filter { it.ct == entry.key }.forEach { it.count = entry.value }
-    }
-    return emptyList.map { CoffeeTypeWithCount(it.ct, it.count) }
-}
