+++ cmp-common/src/jvmMain/kotlin/main.desktop.kt
@@ -1,7 +1,3 @@
-import androidx.compose.foundation.LocalScrollbarStyle
-import androidx.compose.foundation.ScrollbarStyle
-import androidx.compose.material.MaterialTheme
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.drawscope.DrawScope
@@ -13,7 +9,6 @@ import com.arkivanov.decompose.DefaultComponentContext
 import com.arkivanov.essenty.lifecycle.LifecycleRegistry
 import org.jetbrains.compose.reload.DevelopmentEntryPoint
 import org.koin.core.context.GlobalContext.startKoin
-import ru.beryukhov.coffeegram.DefaultPreview
 import ru.beryukhov.coffeegram.appModule
 import ru.beryukhov.coffeegram.components.DefaultRootComponent
 import ru.beryukhov.coffeegram.screens.RootScreen
@@ -42,27 +37,6 @@ fun main() {
     ) {
         DevelopmentEntryPoint {
             RootScreen(root)
-        }
-    }
-}
-
-fun mainOld() = singleWindowApplication(
-    title = "Coffeegram",
-    state = WindowState(width = 800.dp, height = 600.dp),
-    icon = TrayIcon
-) {
-    DevelopmentEntryPoint {
-        CompositionLocalProvider(
-            LocalScrollbarStyle provides ScrollbarStyle(
-                minimalHeight = 16.dp,
-                thickness = 8.dp,
-                shape = MaterialTheme.shapes.small,
-                hoverDurationMillis = 300,
-                unhoverColor = MaterialTheme.colors.onSurface.copy(alpha = 0.12f),
-                hoverColor = MaterialTheme.colors.onSurface.copy(alpha = 0.50f)
-            )
-        ) {
-            DefaultPreview(koinApp.get())
         }
     }
 }
