+++ cmp-common/src/notWasm/kotlin/ru/beryukhov/coffeegram/DI.kt
@@ -3,7 +3,6 @@ package ru.beryukhov.coffeegram
 import org.koin.dsl.module
 import ru.beryukhov.coffeegram.model.DaysCoffeesStore
 import ru.beryukhov.coffeegram.model.DaysCoffeesStoreImpl
-import ru.beryukhov.coffeegram.model.NavigationStore
 import ru.beryukhov.coffeegram.model.ThemeState
 import ru.beryukhov.coffeegram.model.ThemeStore
 import ru.beryukhov.coffeegram.repository.CoffeeStorage
@@ -24,7 +23,4 @@ val appModule = module {
     }
     single<DaysCoffeesStore> { DaysCoffeesStoreImpl(coffeeStorage = get()) }
     single { CoffeeStorage(repository = get()) }
-
-    single { NavigationStore() }
-    single { Dependencies(get(), get(), get()) }
 }
